Extract shared engine test support
