+++ crates/truthdb-core/src/engine/tests.rs
@@ -1,6 +1,6 @@
 use std::path::{Path, PathBuf};
 
-use super::test_support::{test_storage_options, unique_temp_path};
+use super::test_support::*;
 use super::*;
 
 /// Stage 2 exit criterion: search events and relational records share
@@ -58,22 +58,6 @@ fn mixed_search_and_relational_wal_replays_in_order() {
     let _ = std::fs::remove_file(path);
 }
 
-/// Extracts one integer column from a SELECT via the SQL envelope.
-fn sql_column_i64(engine: &Engine, sql: &str, column: usize) -> Vec<i64> {
-    let response = engine.execute(sql).expect("sql");
-    let response: Value = serde_json::from_str(&response).expect("json");
-    assert_eq!(
-        response["kind"], "sql",
-        "expected a rows envelope: {response}"
-    );
-    response["results"][0]["rows"]
-        .as_array()
-        .expect("rows array")
-        .iter()
-        .map(|row| row[column].as_str().expect("cell").parse().expect("i64"))
-        .collect()
-}
-
 #[test]
 fn engine_replay_ignores_relational_wal_records() {
     let path = unique_temp_path("rel-coexistence");
@@ -107,67 +91,6 @@ fn engine_replay_ignores_relational_wal_records() {
     let _ = extent;
 
     let _ = std::fs::remove_file(path);
-}
-
-/// Runs SQL and returns the parsed envelope.
-fn sql(engine: &Engine, text: &str) -> Value {
-    let response = engine.execute(text).expect("execute");
-    serde_json::from_str(&response).expect("json envelope")
-}
-
-/// Runs SQL expected to error and returns the SQL error number from the
-/// envelope's trailing `error`.
-fn sql_error_number(engine: &Engine, text: &str) -> i64 {
-    let env = sql(engine, text);
-    env["error"]["number"]
-        .as_i64()
-        .unwrap_or_else(|| panic!("expected an error envelope, got {env}"))
-}
-
-/// Runs a single-statement SELECT and returns its (columns, rows) where
-/// each cell is `Option<String>` (None = NULL).
-fn sql_rows(engine: &Engine, text: &str) -> (Vec<String>, Vec<Vec<Option<String>>>) {
-    let env = sql(engine, text);
-    assert_eq!(env["kind"], "sql", "expected rows, got {env}");
-    let result = &env["results"][0];
-    assert_eq!(result["type"], "rows", "expected a rowset, got {result}");
-    let columns = result["columns"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .map(|c| c.as_str().unwrap().to_string())
-        .collect();
-    let rows = result["rows"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .map(|row| {
-            row.as_array()
-                .unwrap()
-                .iter()
-                .map(|cell| cell.as_str().map(str::to_string))
-                .collect()
-        })
-        .collect();
-    (columns, rows)
-}
-
-fn new_engine(path: &Path) -> Engine {
-    let storage = Storage::create(path.to_path_buf(), test_storage_options()).expect("create");
-    Engine::new(storage).expect("engine")
-}
-
-/// A table's catalog object id (via `sys.tables`).
-fn table_object_id(engine: &Engine, name: &str) -> u32 {
-    let (_, rows) = sql_rows(
-        engine,
-        &format!("SELECT object_id FROM sys.tables WHERE name = '{name}'"),
-    );
-    rows[0][0]
-        .as_ref()
-        .expect("object_id")
-        .parse()
-        .expect("u32")
 }
 
 #[test]
@@ -5208,12 +5131,6 @@ fn sql_batch_keeps_earlier_results_before_an_error() {
 
 use crate::engine::{BatchOutcome, StatementResult, TxnContext};
 use crate::relstore::types::Datum;
-
-/// Runs a SQL batch through the session path with a persistent transaction
-/// context (as a TDS connection would), returning the typed outcome.
-fn batch(engine: &Engine, ctx: &mut TxnContext, sql: &str) -> BatchOutcome {
-    engine.sql_batch(sql, ctx).expect("sql_batch")
-}
 
 /// The integer `id` column (column 0) of the first rowset in an outcome.
 fn ids(outcome: &BatchOutcome) -> Vec<i32> {
